A lot of work around the discounting logic (fixes, refactoring and unit tests).
The logic that finds the best discount now considers the exchange rates when multiple currencies are involved.

So, if a promotion is in USD but a customer is buying in EUR, we will (a) ignore the promotion if no exchange rate is found or (b) consider the promotion if there *is* an exchange rate.

When finding the best promotion (deepest discount) we resolve the discounts on an order line to the currency on the order's front-line price before choosing the one that provides the deepest discount to the customer.